Fix binary compatibility
