+++ subprojects/distributions/src/changes/accepted-public-api-changes.json
@@ -15,6 +15,73 @@
             "changes": [
                 "org.gradle.api.artifacts.repositories.MetadataSupplierAware"
             ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.eclipse.model.EclipseClasspath",
+            "member": "Constructor EclipseClasspath",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.eclipse.model.EclipseModel",
+            "member": "Constructor EclipseModel",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.eclipse.model.EclipseWtp",
+            "member": "Constructor EclipseWtp",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.idea.model.IdeaModel",
+            "member": "Constructor IdeaModel",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.idea.model.IdeaModule",
+            "member": "Constructor IdeaModule",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.idea.model.IdeaModule",
+            "member": "Constructor org.gradle.plugins.ide.idea.model.IdeaModule(org.gradle.api.Project,org.gradle.api.provider.Provider)",
+            "acceptation": "Users should not instantiate this type"
+        },
+        {
+            "type": "org.gradle.plugins.ide.idea.model.IdeaProject",
+            "member": "Constructor IdeaProject",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
+        },
+        {
+            "type": "org.gradle.plugins.ide.idea.model.IdeaProject",
+            "member": "Constructor org.gradle.plugins.ide.idea.model.IdeaProject(org.gradle.api.Project,org.gradle.api.provider.Provider)",
+            "acceptation": "Users should not instantiate this type"
+        },
+        {
+            "type": "org.gradle.plugins.ide.idea.model.IdeaWorkspace",
+            "member": "Constructor IdeaWorkspace",
+            "acceptation": "Users should not instantiate this type",
+            "changes": [
+                "Constructor has been removed"
+            ]
         }
+
     ]
 }
